refactor : cart page
Related to : #134
