+++ src/components/cart/OrderButton.tsx
@@ -1,0 +1,54 @@
+import { useUser } from '@/hooks/useUser';
+import { cartAtom } from '@/recoil/atom/cartAtom';
+import styles from '@/styles/cart/orderButton.module.css';
+import { CartItemType } from '@/types/bookType';
+import Link from 'next/link';
+import { useRouter } from 'next/navigation';
+import { useSetRecoilState } from 'recoil';
+
+interface OrderButtonProps {
+	checkedItem: string[];
+}
+
+export default function OrderButton({ checkedItem }: OrderButtonProps) {
+	// useRouter 호출
+	const router = useRouter();
+	// 카트 아이템 setState
+	const setCart = useSetRecoilState<CartItemType[]>(cartAtom);
+	// useUser에서 호출한 로그인 상태
+	const { isLoggedIn } = useUser();
+
+	// 주문하기 버튼 클릭 시 동작하는 함수
+	const handleOrderButtonClick = () => {
+		// 로그인 상태가 아닐 경우
+		if (!isLoggedIn) {
+			if (confirm('로그인이 필요한 서비스입니다 로그인 하시겠습니까?')) {
+				// 로그인 페이지로 이동
+				router.push('/login');
+			}
+			// 취소 시 함수 종료
+			return;
+		}
+		// 체크된 아이템이 없을 경우
+		if (checkedItem.length === 0) {
+			alert('주문할 상품을 선택해주세요.');
+			return; // 함수를 여기서 종료
+		}
+		// 주문 확인
+		if (confirm('상품을 주문하시겠습니까?')) {
+			// cart 초기화
+			setCart([]);
+			// 주문 완료 페이지로 이동
+			router.push('/orderComplete');
+		}
+	};
+
+	return (
+		<div className={styles.orderButtonWrap}>
+			<Link href={'/'}>
+				<button>상품 추가</button>
+			</Link>
+			<button onClick={handleOrderButtonClick}>결제하기</button>
+		</div>
+	);
+}
