+++ src/styles/cart/cartFooter.module.css
@@ -1,0 +1,34 @@
+/* cart footer */
+.cartFooter {
+	height: 40px;
+
+	border-bottom: 1px solid #a9a9a9;
+}
+.cartDeleteBtn {
+	width: 59px;
+	height: inherit;
+	font-size: 12px;
+
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+.cartDeleteBtn > button {
+	width: 50px;
+	height: 20px;
+	border: 1px solid black;
+	border-radius: 50px;
+	font-size: 11px;
+
+	border: none;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	cursor: pointer;
+
+	background-color: #f8f8f8;
+}
+.cartDeleteBtn > button:hover {
+	color: #ffffff;
+	background-color: #231304;
+}
